Rewrite behavior of contentTypeHint attribute.
This change settles on "" as default or unset values for the attribute.
It separates and defines specific behaviors for both setting and getting
values, and describes how to behave on invalid values. This fixes #2,
fixes #3 and fixes #4.

Additionally this adds a table for the "" default value which is valid
for both audio and video tracks. There was a consideration to include ""
in both enumerations, and then use an union between the two, but it's
unclear whether unions with intersecting values are permitted.