Fix links to Audio/Video Content Hints sections
Presently ReSpec believes that there should be a link to the "Audio Content Hints" 
and "Video Content Hints" sections because they have a `<h3>` . So it renders a § next
to it. Unfortunately, that § produces a URL fragment that is inherited from the previous
section, and so using the fragment to deep-link to these sections results in the user being
directed to the wrong section (the parent section element).

By adding a section wrapper around the `<h3>` headers, ReSpec auto-generates an ID based
on the `<h3>`'s contents, when enables the § to work propertly.

All the other changes are just indentation adjustments based on the added section elements.